Move fill utility to header file
