+++ include/boost/histogram/python/axis.hpp
@@ -4,6 +4,8 @@
 // (See accompanying file LICENSE_1_0.txt
 // or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#pragma once
+
 #include <boost/histogram.hpp>
 
 #include <tuple>
