+++ include/boost/histogram/python/storage.hpp
@@ -1,3 +1,9 @@
+// Copyright 2018 Henry Schreiner
+//
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt
+// or copy at http://www.boost.org/LICENSE_1_0.txt)
+
 #pragma once
 
 #include <boost/histogram.hpp>
